fix(test): TestRedisConfig 분리
